Add files via upload
This function uses a peak detector to calculate the wrapper of a given signal.

Whenever there's a change in the polarity of the input signal, the peak detector is reset to start searching for the peak value among polarity switchings. When that happens, the amplitude and the time stamp in which it occurred are recorded. 
In the end a vector containing the peaks amplitude and another vector containing the time stamps of the peaks are returned.